separate gemmi and mc tests into separate suites
